adding an exception for the case when sample_mask is empty
